+++ packages/account-message-bridge/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sophon-labs/account-message-bridge
 
+## 3.0.0-next.10
+
+### Patch Changes
+
+- More languages support
+
+## 3.0.0-next.9
+
+### Patch Changes
+
+- Fixing issue on consent
+
+## 3.0.0-next.8
+
+### Patch Changes
+
+- Consent flag
+
 ## 3.0.0-next.7
 
 ### Patch Changes