+++ packages/account-partner/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sophon-labs/account-partner
 
+## 3.0.0-next.10
+
+### Patch Changes
+
+- More languages support
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.10
+
+## 3.0.0-next.9
+
+### Patch Changes
+
+- Fixing issue on consent
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.9
+
+## 3.0.0-next.8
+
+### Patch Changes
+
+- Consent flag
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.8
+
 ## 3.0.0-next.7
 
 ### Patch Changes